// src/app/shared/interfaces/storage.interface.ts
// Generated by https://quicktype.io

export interface ListStorage {
  result: Result;
}

export interface Result {
  itemsList: Stogare[];
  total: number;
  offset: number;
  limit: number;
  totalPages: number;
  page: number;
  pagingCounter: number;
  hasPrevPage: boolean;
  hasNextPage: boolean;
  prevPage: null;
  nextPage: null;
}

export interface Stogare {
  _id: string;
  name: string;
  identificador: string;
}