// src/app/shared/interfaces/proveedores.interface.ts
// Generated by https://quicktype.io

export interface ListProveedores {
  result: Result;
}

export interface Result {
  itemsList: Proveedor[];
  total: number;
  offset: number;
  limit: number;
  totalPages: number;
  page: number;
  pagingCounter: number;
  hasPrevPage: boolean;
  hasNextPage: boolean;
  prevPage: null;
  nextPage: null;
}

export interface Proveedor {
  _id: string;
  name: string;
  country?: string;
  address?: string;
  phone?: number;
  contact?: string;
  observations?: string;
}